refactor: extract formatCategory to shared util and MarkPips to own CSS module
- Extract duplicated formatCategory + CATEGORY_LABELS to
  src/utils/formatting.ts — both AccessoriesExplorer and
  [slug].astro now import from the shared location (closes #318)
- Move MarkPips styles from GenreGuide.module.css to
  shared/MarkPips.module.css — MarkPips no longer imports a
  sibling's CSS module directly (closes #319)
- Remove dead .sweetSpot CSS class from GenreGuide.module.css
- Add formatCategory tests to formatting.test.ts
- Net: +38 −102 lines, 175 tests passing
